requirements.txt
cloudinary
requests
python-dotenv